arrow 클래스 생성

+++ DX_1800/DX_1800/Framework/Utility/Timer.cpp
@@ -10,7 +10,7 @@ Timer::Timer()
 	// 프로그램이 시작할 때 지금까지 CPU의 진동수
 	QueryPerformanceCounter((LARGE_INTEGER*) &_lastTime);
 
-	_timeScale = 1.0 / (double)_periodFrequency;
+	_timeScale = 1.0f / (float)_periodFrequency;
 }
 
 Timer::~Timer()
@@ -20,14 +20,14 @@ Timer::~Timer()
 void Timer::Update()
 {
 	QueryPerformanceCounter((LARGE_INTEGER*) &_curTime);
-	_deltaTime = (double)(_curTime - _lastTime) * _timeScale;
+	_deltaTime = (float)(_curTime - _lastTime) * _timeScale;
 
 	if (_lockFPS != 0)
 	{
-		while (_deltaTime < (1.0 / _lockFPS))
+		while (_deltaTime < (1.0f / _lockFPS))
 		{
 			QueryPerformanceCounter((LARGE_INTEGER*)&_curTime);
-			_deltaTime = (double)(_curTime - _lastTime) * _timeScale;
+			_deltaTime = (float)(_curTime - _lastTime) * _timeScale;
 		}
 	}
