+++ DX_1800/DX_1800/Framework/Utility/Timer.h
@@ -29,15 +29,15 @@ public:
 	void Update();
 
 	int GetFPS() { return _frameRate; }
-	void LockFPS(double value) { _lockFPS = value; }
-	double GetDeltaTime() { return _deltaTime; }
-	double GetRunTime() { return _runTime; }
+	void LockFPS(float value) { _lockFPS = value; }
+	float GetDeltaTime() { return _deltaTime; }
+	float GetRunTime() { return _runTime; }
 
 private:
 	static Timer* _instance;
 
-	double _timeScale = 0.0;
-	double _deltaTime = 0.0; // 한번 업데이트에 걸리는 시간
+	float _timeScale = 0.0;
+	float _deltaTime = 0.0; // 한번 업데이트에 걸리는 시간
 
 	unsigned __int64 _curTime = 0;
 	unsigned __int64 _lastTime = 0;
@@ -46,8 +46,8 @@ private:
 	unsigned int _frameRate = 0;
 	unsigned int _frameCount = 0;
 
-	double _oneSecCount = 0.0;
-	double _runTime = 0.0;
-	double _lockFPS = 0.0;
+	float _oneSecCount = 0.0;
+	float _runTime = 0.0;
+	float _lockFPS = 0.0;
 };
